Add files via upload
I am incredibly grateful for this remarkable achievement as I was one of the finalist in Google Girl Hackathon 2023. This year the Hackathon comprised 3 learning cohorts, and I had the privilege of being part of the networking and troubleshooting cohort.

During the Ideathon Finale round, I had the opportunity to present my idea on designing a highly secure, reliable, and scalable fintech application. Throughout the process, I gained invaluable knowledge and skills that have expanded my horizons. This describes my journey so far.